Quresh_Database/requirements.txt
Flask
Flask-SQLAlchemy
pandas
reportlab
openpyxl
Pillow
beautifulsoup4
